services/collector/requirements.txt
python-telegram-bot>=20.7
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tzdata>=2024.1
